Add running as node in heroku

+++ www/js/controllers/reed.js
@@ -1,4 +1,4 @@
-app.controller('ReedCtrl', function($scope, $state, $timeout, reed, reedsService) {
+app.controller('ReedCtrl', function($scope, $state, $ionicPopup, $timeout, reed, reedsService) {
     var updateTimer = false;
 
     $scope.reed = reed;
@@ -7,15 +7,23 @@ app.controller('ReedCtrl', function($scope, $state, $timeout, reed, reedsService
         if (updateTimer) $timeout.cancel(updateTimer);
 
         updateTimer = $timeout(function() {
-          reedsService.update($scope.reed._id, $scope.reed);
+            reedsService.update($scope.reed._id, $scope.reed);
         }, 250);
     };
 
     $scope.delete = function() {
-        reedsService.delete($scope.reed._id).success(function() {
-            $state.go('app.reeds.index', {}, {
-                reload: true
-            });
+        var confirmPopup = $ionicPopup.confirm({
+            title: 'Delete',
+            template: 'Are you sure you want to delete this reed?'
+        });
+        confirmPopup.then(function(res) {
+            if (res) {
+                reedsService.delete($scope.reed._id).success(function() {
+                    $state.go('app.reeds.index', {}, {
+                        reload: true
+                    });
+                });
+            }
         });
     };
 });
